+++ public/pages/about.html
@@ -9,7 +9,7 @@
 	<main>
         <div id="navbar">
             <span id="navbarSpacer"></span>
-            <span id="pageTitle">About</span>
+            <span class="title" id="pageTitle">About</span>
             <span id="logo">jspuyo</span>
         </div>
         <div id="text-content">
@@ -36,7 +36,7 @@
         <footer>
             Back to <a href="/">jspuyo</a>.
         </footer>
-        <script src="../limited.js"></script>
+        <script src="../bundle.js"></script>
     </main>
 </body>
 </html>
