+++ public/pages/gallery.html
@@ -13,7 +13,7 @@
 	<main>
         <div id="navbar">
             <span id="navbarSpacer"></span>
-            <span id="pageTitle">Gallery</span>
+            <span class="title" id="pageTitle">Gallery</span>
             <span id="logo">jspuyo</span>
         </div>
         <div id="gamesTracked"></div>
@@ -66,7 +66,7 @@
         <footer>
             Back to <a href="/">jspuyo</a>.
         </footer>
-        <script src="../limited.js"></script>
+        <script src="../bundle.js"></script>
     </main>
 </body>
 </html>
