+++ public/pages/privacy.html
@@ -3,13 +3,13 @@
     <title>jspuyo</title>
     <meta charset="utf-8" />
     <link href="https://fonts.googleapis.com/css2?family=Chelsea+Market&family=Rock+Salt&display=swap" rel="stylesheet">
-    <link rel="stylesheet" type="text/css" href="./../style.css" />
+    <link rel="stylesheet" type="text/css" href="style.css" />
 </head>
 <body>
 	<main>
         <div id="navbar">
             <span id="navbarSpacer"></span>
-            <span id="pageTitle">Privacy Policy</span>
+            <span class="title" id="pageTitle">Privacy Policy</span>
             <span id="logo">jspuyo</span>
         </div>
         <div id=text-content>
@@ -27,7 +27,7 @@
 
 			<p>We reserve the right to change this Privacy Policy at any time, and will update the “Last Updated” date at the beginning of this document whenever it is changed. You are encouraged to review this document periodically to stay up to date.</p>
 		</div>
-		<script src="../limited.js"></script>
+		<script src="../bundle.js"></script>
     </main>
 </body>
 </html>
