+++ public/pages/terms.html
@@ -9,13 +9,13 @@
 	<main>
         <div id="navbar">
             <span id="navbarSpacer"></span>
-            <span id="pageTitle">Terms of Service</span>
+            <span class="title" id="pageTitle">Terms of Service</span>
             <span id="logo">jspuyo</span>
         </div>
         <div id=text-content>
 			<p>This page is still under construction. Check back later!</p>
 		</div>
-		<script src="../limited.js"></script>
+		<script src="../bundle.js"></script>
     </main>
 </body>
 </html>
