+++ src/_css/pages/_base_pages.scss
@@ -1,0 +1,30 @@
+@use "../base";
+@use "../modal";
+@use "charts";
+@use "guide";
+
+a {
+	color: rgb(150, 150, 240);
+}
+
+.title {
+	color: white;
+	font: 40px base.$font-title;
+	text-shadow: base.$text-drop-shadow;
+}
+
+// Titles for other pages
+#pageTitle {
+	padding: 10px 0px 10px 40px;
+}
+
+#text-content {
+	margin: 3%;
+	padding: 1em;
+
+	color: white;
+	font: 20px base.$font-title;
+
+	background: rgba(40, 40, 40, 0.8);
+	border: 2px solid white;
+}
