+++ src/_css/pages/_guide.scss
@@ -1,0 +1,77 @@
+@use "../base";
+@use "../modal";
+
+%guide-button {
+	padding: 4px 20px;
+
+	font-size: 1.45vw;
+}
+
+#wrapper {
+	display: flex;
+	flex: 0 1 auto;
+	justify-content: center;
+	width: 100%;
+	height: 100%;
+	margin: auto 0;
+}
+
+#simulator {
+	width: 352;
+	height: 640;
+	margin: auto 0;
+}
+
+#content {
+	width: 50%;
+	margin: auto 0 auto 50px;
+}
+
+#guideTitle {
+	font-size: 60px;
+	text-align: center;
+}
+
+#info {
+	padding: 10px 30px;
+
+	color: white;
+	font-size: 1.4vw;
+
+	background: rgba(40, 40, 40, 0.8);
+	border: 4px solid white;
+}
+
+#navigation {
+	display: flex;
+	justify-content: space-evenly;
+	margin: 40px 0;
+}
+
+#settingsButton {
+	@extend %guide-button;
+	display: block;
+	margin: 0 auto 20px;
+
+	background-color: #6945b3;
+}
+
+#simulatorStart {
+	@extend %guide-button;
+	background-color: base.$dark-green;
+}
+
+#simulatorStop {
+	@extend %guide-button;
+	background-color: base.$error-colour;
+}
+
+#previousButton {
+	@extend %guide-button;
+	background-color: #791639;
+}
+
+#nextButton {
+	@extend %guide-button;
+	background-color: #165579;
+}
