+++ src/_css/style.scss
@@ -3,5 +3,5 @@
 @use "navbar";
 @use "mainpage";
 @use "dialog";
-@use "pages";
+@use "pages/base_pages";
 @use "firebase";
